+++ test/smoke/v7/default.rb
@@ -6,36 +6,36 @@
 # found at http://inspec.io/docs/reference/resources/
 
 if %w(redhat fedora amazon).include?(os[:family])
-  describe file('/etc/yum.repos.d/elastic6.repo') do
-    its('content') { should match %r{https://artifacts.elastic.co/packages/6.x/yum} }
+  describe file('/etc/yum.repos.d/elastic7.repo') do
+    its('content') { should match %r{https://artifacts.elastic.co/packages/7.x/yum} }
   end
 else
-  describe file('/etc/apt/sources.list.d/elastic6.list') do
-    its('content') { should match %r{https://artifacts.elastic.co/packages/6.x/apt} }
+  describe file('/etc/apt/sources.list.d/elastic7.list') do
+    its('content') { should match %r{https://artifacts.elastic.co/packages/7.x/apt} }
   end
 end
 
 describe package('filebeat') do
   it { should be_installed }
-  its('version') { should match '6.8.8' }
+  its('version') { should match '7.6.2' }
 end
 
 describe package('packetbeat') do
   it { should be_installed }
-  its('version') { should match '6.8.8' }
+  its('version') { should match '7.6.2' }
 end
 
 describe package('metricbeat') do
   it { should be_installed }
-  its('version') { should match '6.8.8' }
+  its('version') { should match '7.6.2' }
 end
 
 describe package('heartbeat-elastic') do
   it { should be_installed }
-  its('version') { should match '6.8.8' }
+  its('version') { should match '7.6.2' }
 end
 
 describe package('auditbeat') do
   it { should be_installed }
-  its('version') { should match '6.8.8' }
+  its('version') { should match '7.6.2' }
 end